Revert "Fix types for 32 bit build"
This reverts commit 1f64e45e0b0580921bb4f71d247f2d002a6b9149.

+++ source/rmg-dispatcher.c
@@ -281,8 +281,8 @@ do_process_service_crash_event (RmgDispatcher *dispatcher, RmgDEvent *event)
   gboolean action_reset_after = false;
 
   g_autoptr (GError) error = NULL;
-  guint64 service_hash = 0;
-  gint64 rvector = 0;
+  gulong service_hash = 0;
+  glong rvector = 0;
 
   g_assert (dispatcher);
   g_assert (event);
@@ -347,10 +347,10 @@ do_process_service_crash_event (RmgDispatcher *dispatcher, RmgDEvent *event)
 
   if (action_type != ACTION_INVALID)
     {
-      g_info ("Action '%s' requiered for service='%s' rvector=%lld",
+      g_info ("Action '%s' requiered for service='%s' rvector=%ld",
               rmg_utils_action_name (action_type),
               event->service_name,
-              (long long int)rvector);
+              rvector);
     }
 
   switch (action_type)
